fix(models): address pull review feedback
