+++ pkg/services/models/asset_scope_characterization_test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 
 	models "github.com/portpowered/infinite-you/pkg/services/models"
+	pullsupport "github.com/portpowered/infinite-you/pkg/services/models/internal/pullsupport"
 )
 
 func (unsupportedRuntimeScopePeer) PrepareModelAssets(
@@ -441,7 +442,7 @@ func TestPullDiagnosticsPreserveSafeFactsWithoutRenderingRawCause(t *testing.T) 
 		RequestURL:         "https://user:password@example.test/owner/repo/resolve/rev-1/weights/model.gguf?download=true&token=secret",
 		UpstreamStatusCode: 502,
 	}
-	safe := models.NewPullDiagnosticsError(diagnostics, raw)
+	safe := pullsupport.NewPullDiagnosticsError(diagnostics, raw)
 	if safe == nil || strings.Contains(safe.Error(), raw.Error()) || strings.Contains(safe.Error(), "secret") ||
 		strings.Contains(safe.Error(), "C:\\private") {
 		t.Fatalf("safe diagnostics = %v, leaked raw cause", safe)
@@ -449,15 +450,15 @@ func TestPullDiagnosticsPreserveSafeFactsWithoutRenderingRawCause(t *testing.T) 
 	if !errors.Is(safe, raw) {
 		t.Fatalf("safe diagnostics = %v, want errors.Is to preserve raw cause identity", safe)
 	}
-	if got := models.PullDiagnosticsFromError(safe); got.RequestURL != "https://example.test/owner/repo/resolve/rev-1/weights/model.gguf?download=true" ||
+	if got := pullsupport.PullDiagnosticsFromError(safe); got.RequestURL != "https://example.test/owner/repo/resolve/rev-1/weights/model.gguf?download=true" ||
 		got.UpstreamStatusCode != http.StatusBadGateway {
 		t.Fatalf("normalized diagnostics = %#v, want redacted URL and status", got)
 	}
 
-	stage := models.WrapPullStage(
+	stage := pullsupport.WrapPullStage(
 		models.PullStageIntegrityVerification, "OMNIVOICE_Q4_K_M", "verify downloaded asset", "weights/model.gguf", raw,
 	)
-	wrapped := models.WrapPullDiagnostics(diagnostics, stage)
+	wrapped := pullsupport.WrapPullDiagnostics(diagnostics, stage)
 	var stageError *models.PullStageError
 	if !errors.As(wrapped, &stageError) || stageError.Stage != models.PullStageIntegrityVerification {
 		t.Fatalf("wrapped diagnostics = %v, want typed stage cause", wrapped)
