+++ pkg/services/models/internal/local/catalog.go
@@ -10,6 +10,7 @@ import (
 
 	models "github.com/portpowered/infinite-you/pkg/services/models"
 	managedruntime "github.com/portpowered/infinite-you/pkg/services/models/internal/managedruntime"
+	pullsupport "github.com/portpowered/infinite-you/pkg/services/models/internal/pullsupport"
 	modelcatalog "github.com/portpowered/infinite-you/pkg/services/models/internal/services/catalog"
 )
 
@@ -764,11 +765,11 @@ func inspectPullCache(
 	}
 	inspection, err := inspector.InspectRuntimeCache(ctx, runtimeCfg, modelName)
 	if err != nil {
-		stage := models.PullStageForError(err)
+		stage := pullsupport.PullStageForError(err)
 		if stage == "" {
 			stage = models.PullStageReadinessEvaluation
 		}
-		return RuntimeCacheInspection{}, models.WrapPullStage(
+		return RuntimeCacheInspection{}, pullsupport.WrapPullStage(
 			stage, modelName, "inspect managed runtime cache", "", err,
 		)
 	}
@@ -787,11 +788,16 @@ func classifiedPullFailure(
 ) (models.PullResult, error) {
 	pullOutcome, readiness := ClassifyPullFailure(cause)
 	result := base
-	diagnostics := models.MergePullDiagnostics(
+	stage := pullsupport.PullStageForError(cause)
+	operation := "classify pull failure"
+	if stage == models.PullStageSourceResolution {
+		operation = "resolve model source"
+	}
+	diagnostics := pullsupport.MergePullDiagnostics(
 		result.PullDiagnostics,
-		models.PullDiagnosticsFromError(cause),
+		pullsupport.PullDiagnosticsFromError(cause),
 	).WithDefaults(
-		modelName, result.SourceID, result.Revision, "", "classify pull failure",
+		modelName, result.SourceID, result.Revision, "", operation,
 	)
 	if strings.TrimSpace(result.ModelName) == "" {
 		result.ModelName = strings.TrimSpace(modelName)
@@ -806,7 +812,7 @@ func classifiedPullFailure(
 	result.SourceKind = strings.TrimSpace(resolution.SourceKind)
 	result.SourceID = strings.TrimSpace(resolution.SourceID)
 	result.ResolverNotes = strings.TrimSpace(resolution.ResolverNotes)
-	result.FailureStage = models.PullStageForError(cause)
+	result.FailureStage = stage
 	if result.FailureStage == "" &&
 		pullOutcome == managedPullOutcomeAssetPreparationFailed &&
 		!errors.Is(cause, context.Canceled) && !errors.Is(cause, context.DeadlineExceeded) {
@@ -875,195 +881,4 @@ func SelectInvocationWorker(
 		return nil, models.RuntimeOperation{}, fmt.Errorf("%w: worker %q for model %q does not support operation %q", modelcatalog.ErrUnsupportedOperation, matchedWorkerName, modelName, operationName)
 	}
 	return nil, models.RuntimeOperation{}, fmt.Errorf("%w: %s", managedruntime.ErrNotFound, modelName)
-}
-
-const (
-	legacyPullOutcomePulled         = "PULLED"
-	legacyPullOutcomeAlreadyPresent = "ALREADY_PRESENT"
-	legacyPullOutcomeFailed         = "FAILED"
-
-	managedPullOutcomeAlreadyReady                = "ALREADY_READY"
-	managedPullOutcomeInstalledSuccessfully       = "INSTALLED_SUCCESSFULLY"
-	managedPullOutcomeAlreadyPresent              = "ALREADY_PRESENT"
-	managedPullOutcomeStillLoading                = "STILL_LOADING"
-	managedPullOutcomeTimedOut                    = "TIMED_OUT"
-	managedPullOutcomeCancelled                   = "CANCELLED"
-	managedPullOutcomeSourceFetchFailed           = "SOURCE_FETCH_FAILED"
-	managedPullOutcomeUnsupportedRuntime          = "UNSUPPORTED_RUNTIME"
-	managedPullOutcomeSourceResolutionFailed      = "SOURCE_RESOLUTION_FAILED"
-	managedPullOutcomeIntegrityVerificationFailed = "INTEGRITY_VERIFICATION_FAILED"
-	managedPullOutcomeAssemblyFailed              = "ASSEMBLY_FAILED"
-	managedPullOutcomeCacheInstallationFailed     = "CACHE_INSTALLATION_FAILED"
-	managedPullOutcomeReadinessEvaluationFailed   = "READINESS_EVALUATION_FAILED"
-	managedPullOutcomeAssetPreparationFailed      = "ASSET_PREPARATION_FAILED"
-
-	managedReadinessReady       = "READY"
-	managedReadinessMissing     = "MISSING"
-	managedReadinessLoading     = "LOADING"
-	managedReadinessFailed      = "FAILED"
-	managedReadinessUnsupported = "UNSUPPORTED"
-
-	managedLifecycleInstalling   = "INSTALLING"
-	managedLifecycleInstalled    = "INSTALLED"
-	managedLifecycleNotInstalled = "NOT_INSTALLED"
-)
-
-// EnrichPullResult projects a service-owned pull result into managed-runtime
-// readiness, lifecycle, and source diagnostics using post-pull cache inspection.
-func EnrichPullResult(
-	result models.PullResult,
-	inspection RuntimeCacheInspection,
-	resolution ManagedRuntimeSourceResolution,
-) models.PullResult {
-	outcome, readiness, lifecycle := classifySuccessfulPull(result, inspection)
-	result.ManagedPullOutcome = outcome
-	result.ReadinessState = readiness
-	result.LifecycleState = lifecycle
-	if readiness == managedReadinessFailed && result.FailureStage == "" {
-		result.FailureStage = pullStageForManagedOutcome(outcome)
-	}
-	result.SourceKind = strings.TrimSpace(resolution.SourceKind)
-	result.SourceID = strings.TrimSpace(resolution.SourceID)
-	result.ResolverNotes = strings.TrimSpace(resolution.ResolverNotes)
-	return result
-}
-
-// ClassifyPullFailure maps pull errors to managed-runtime pull outcomes and
-// readiness states for logging, metrics, and stable customer-facing vocabulary.
-func ClassifyPullFailure(err error) (pullOutcome string, readiness string) {
-	if err == nil {
-		return "", ""
-	}
-	switch {
-	case errors.Is(err, context.DeadlineExceeded):
-		return managedPullOutcomeTimedOut, managedReadinessFailed
-	case errors.Is(err, context.Canceled), errors.Is(err, models.ErrAssetCancelled):
-		return managedPullOutcomeCancelled, managedReadinessFailed
-	case errors.Is(err, models.ErrPullUnsupported):
-		return managedPullOutcomeUnsupportedRuntime, managedReadinessUnsupported
-	}
-	switch models.PullStageForError(err) {
-	case models.PullStageSourceResolution:
-		return managedPullOutcomeSourceResolutionFailed, managedReadinessFailed
-	case models.PullStageSourceFetch:
-		return managedPullOutcomeSourceFetchFailed, managedReadinessFailed
-	case models.PullStageIntegrityVerification:
-		return managedPullOutcomeIntegrityVerificationFailed, managedReadinessFailed
-	case models.PullStageAssembly:
-		return managedPullOutcomeAssemblyFailed, managedReadinessFailed
-	case models.PullStageCacheInstallation:
-		return managedPullOutcomeCacheInstallationFailed, managedReadinessFailed
-	case models.PullStageReadinessEvaluation:
-		return managedPullOutcomeReadinessEvaluationFailed, managedReadinessFailed
-	}
-	if isIntegrityFailureMessage(err.Error()) {
-		return managedPullOutcomeIntegrityVerificationFailed, managedReadinessFailed
-	}
-	if isSourceFetchFailureMessage(err.Error()) {
-		return managedPullOutcomeSourceFetchFailed, managedReadinessFailed
-	}
-	return managedPullOutcomeAssetPreparationFailed, managedReadinessFailed
-}
-
-func classifySuccessfulPull(result models.PullResult, inspection RuntimeCacheInspection) (pullOutcome, readiness, lifecycle string) {
-	legacyOutcome := strings.ToUpper(strings.TrimSpace(result.Outcome))
-	switch legacyOutcome {
-	case legacyPullOutcomePulled:
-		pullOutcome = managedPullOutcomeInstalledSuccessfully
-	case legacyPullOutcomeAlreadyPresent:
-		pullOutcome = managedPullOutcomeAlreadyPresent
-	default:
-		pullOutcome = managedPullOutcomeUnsupportedRuntime
-	}
-
-	if inspection.Supported {
-		projection := managedruntime.ProjectManagedRuntimeState(
-			managedRuntimeCacheFacts(managedruntime.LocalityLocal, inspection),
-			models.ManagedRuntimeHostFacts{},
-		)
-		readiness = string(projection.ReadinessState)
-		lifecycle = string(projection.LifecycleState)
-		switch projection.ReadinessState {
-		case managedruntime.ReadinessStateReady:
-			if pullOutcome == managedPullOutcomeAlreadyPresent {
-				pullOutcome = managedPullOutcomeAlreadyReady
-			}
-		case managedruntime.ReadinessStateLoading:
-			pullOutcome = managedPullOutcomeStillLoading
-		case managedruntime.ReadinessStateFailed:
-			pullOutcome = pullOutcomeForFailedInspection(inspection)
-		case managedruntime.ReadinessStateMissing:
-			if pullOutcome == managedPullOutcomeInstalledSuccessfully {
-				readiness = managedReadinessLoading
-				lifecycle = managedLifecycleInstalling
-				pullOutcome = managedPullOutcomeStillLoading
-			}
-		}
-		return pullOutcome, readiness, lifecycle
-	}
-
-	readiness = managedReadinessReady
-	lifecycle = managedLifecycleInstalled
-	if pullOutcome == managedPullOutcomeAlreadyPresent {
-		pullOutcome = managedPullOutcomeAlreadyReady
-	}
-	return pullOutcome, readiness, lifecycle
-}
-
-func pullOutcomeForFailedInspection(inspection RuntimeCacheInspection) string {
-	reason := strings.ToLower(strings.TrimSpace(inspection.FailureReason))
-	if strings.Contains(reason, "integrity") || strings.Contains(reason, "checksum") ||
-		strings.Contains(reason, "unexpected size") {
-		return managedPullOutcomeIntegrityVerificationFailed
-	}
-	return managedPullOutcomeReadinessEvaluationFailed
-}
-
-func pullStageForManagedOutcome(outcome string) models.PullStage {
-	switch outcome {
-	case managedPullOutcomeSourceResolutionFailed:
-		return models.PullStageSourceResolution
-	case managedPullOutcomeSourceFetchFailed:
-		return models.PullStageSourceFetch
-	case managedPullOutcomeIntegrityVerificationFailed:
-		return models.PullStageIntegrityVerification
-	case managedPullOutcomeAssemblyFailed:
-		return models.PullStageAssembly
-	case managedPullOutcomeCacheInstallationFailed:
-		return models.PullStageCacheInstallation
-	case managedPullOutcomeReadinessEvaluationFailed:
-		return models.PullStageReadinessEvaluation
-	default:
-		return ""
-	}
-}
-
-func isIntegrityFailureMessage(message string) bool {
-	trimmed := strings.ToLower(strings.TrimSpace(message))
-	if trimmed == "" {
-		return false
-	}
-	for _, fragment := range []string{"integrity", "checksum", "unexpected size"} {
-		if strings.Contains(trimmed, fragment) {
-			return true
-		}
-	}
-	return false
-}
-
-func isSourceFetchFailureMessage(message string) bool {
-	trimmed := strings.ToLower(strings.TrimSpace(message))
-	if trimmed == "" {
-		return false
-	}
-	for _, fragment := range []string{
-		"pull model manifest",
-		"download model asset",
-		"model asset request failed",
-	} {
-		if strings.Contains(trimmed, fragment) {
-			return true
-		}
-	}
-	return false
 }
