+++ pkg/services/models/internal/services/assets/internal/service/generic.go
@@ -2,19 +2,17 @@ package service
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
 	"net/url"
-	"path"
 	"path/filepath"
 	"sort"
 	"strings"
 
 	models "github.com/portpowered/infinite-you/pkg/services/models"
+	pullsupport "github.com/portpowered/infinite-you/pkg/services/models/internal/pullsupport"
 	scopedassets "github.com/portpowered/infinite-you/pkg/services/models/internal/services/assets"
 )
 
@@ -172,7 +170,10 @@ func (s *service) genericPreparationPlan(
 	}
 	source, err := s.resolveGenericSource(ctx, scope, rawReference)
 	if err != nil {
-		return genericPreparationPlan{}, err
+		return genericPreparationPlan{}, pullsupport.WrapPullStage(
+			models.PullStageSourceResolution, request.Name,
+			"resolve model source", rawReference, err,
+		)
 	}
 	source.modelName = request.Name
 	modelRequirements, err := s.genericModelRequirements(ctx, source, request.Artifacts)
@@ -212,7 +213,10 @@ func (s *service) genericBackendPlan(
 		var err error
 		backendSource, err = s.resolveGenericSource(ctx, scope, request.BackendReference.NameOrURI)
 		if err != nil {
-			return genericSource{}, nil, nil, err
+			return genericSource{}, nil, nil, pullsupport.WrapPullStage(
+				models.PullStageSourceResolution, request.Name,
+				"resolve backend source", request.BackendReference.NameOrURI, err,
+			)
 		}
 	}
 	if backend := strings.TrimSpace(request.Backend); backend != "" {
@@ -565,7 +569,7 @@ func (s *service) fetchGenericManifest(
 	}
 	request, err := http.NewRequestWithContext(ctx, http.MethodGet, requestURL, nil)
 	if err != nil {
-		return nil, models.WrapPullDiagnostics(
+		return nil, pullsupport.WrapPullDiagnostics(
 			diagnostics,
 			fmt.Errorf("%w: model manifest request is invalid", models.ErrSourceFetchFailed),
 		)
@@ -575,7 +579,7 @@ func (s *service) fetchGenericManifest(
 		if contextErr := assetContextError(ctx); contextErr != nil {
 			return nil, contextErr
 		}
-		return nil, models.WrapPullDiagnostics(
+		return nil, pullsupport.WrapPullDiagnostics(
 			diagnostics,
 			fmt.Errorf("%w: model manifest fetch failed", models.ErrSourceFetchFailed),
 		)
@@ -583,7 +587,7 @@ func (s *service) fetchGenericManifest(
 	defer response.Body.Close()
 	if response.StatusCode != http.StatusOK {
 		diagnostics.UpstreamStatusCode = response.StatusCode
-		return nil, models.WrapPullDiagnostics(
+		return nil, pullsupport.WrapPullDiagnostics(
 			diagnostics,
 			fmt.Errorf("%w: model manifest fetch failed", models.ErrSourceFetchFailed),
 		)
@@ -591,7 +595,7 @@ func (s *service) fetchGenericManifest(
 	var payload upstreamModel
 	if err := json.NewDecoder(response.Body).Decode(&payload); err != nil {
 		diagnostics.Operation = "decode model manifest"
-		return nil, models.WrapPullDiagnostics(
+		return nil, pullsupport.WrapPullDiagnostics(
 			diagnostics,
 			fmt.Errorf("%w: model manifest is invalid", models.ErrSourceFetchFailed),
 		)
@@ -619,9 +623,13 @@ func (s *service) fetchGenericManifest(
 			missing := diagnostics
 			missing.File = name
 			missing.Operation = "resolve manifest artifact"
-			return nil, models.WrapPullDiagnostics(
+			return nil, pullsupport.WrapPullDiagnostics(
 				missing,
-				fmt.Errorf("%w: model manifest is missing an artifact", models.ErrSourceFetchFailed),
+				pullsupport.WrapPullStage(
+					models.PullStageSourceResolution, source.modelName,
+					missing.Operation, name,
+					fmt.Errorf("%w: model manifest is missing an artifact", models.ErrModelReferenceUnknown),
+				),
 			)
 		}
 		size := sibling.Size
@@ -679,7 +687,7 @@ func mergeGenericManifest(
 		}
 		return nil, fmt.Errorf(
 			"%w: model manifest is missing asset %q",
-			models.ErrSourceFetchFailed, artifact.requirement.Name,
+			models.ErrModelReferenceUnknown, artifact.requirement.Name,
 		)
 	}
 	return result, nil
@@ -754,250 +762,4 @@ func genericOverlay(
 	}
 	sort.Strings(names)
 	return names[0], overlays[names[0]].Clone(), true
-}
-
-func parseGenericSource(raw string) (genericSource, error) {
-	lower := strings.ToLower(strings.TrimSpace(raw))
-	switch {
-	case strings.HasPrefix(lower, "hf://"):
-		return parseGenericHFSource(raw)
-	case strings.HasPrefix(lower, "file://"):
-		return parseGenericFileSource(raw)
-	case strings.HasPrefix(lower, "https://"):
-		return parseGenericReleaseSource(raw)
-	case strings.Contains(raw, "://"):
-		return genericSource{}, models.ErrAssetSourceUnsupported
-	case looksLikeLocalPath(raw):
-		return genericSource{kind: genericSourceLocal, safe: "local://path", localPath: raw}, nil
-	default:
-		return genericSource{}, models.ErrAssetSourceUnsupported
-	}
-}
-
-func parseGenericReleaseSource(raw string) (genericSource, error) {
-	parsed, err := url.Parse(raw)
-	if err != nil || parsed.Scheme != "https" || parsed.Host != "github.com" ||
-		parsed.User != nil || parsed.RawQuery != "" || parsed.Fragment != "" ||
-		!strings.Contains(path.Clean(parsed.Path), "/releases/download/") {
-		return genericSource{}, models.ErrModelReferenceInvalid
-	}
-	name := path.Base(parsed.Path)
-	if name == "." || name == "/" || strings.TrimSpace(name) == "" || strings.ContainsAny(name, "\\\x00") {
-		return genericSource{}, models.ErrModelReferenceInvalid
-	}
-	checksum := sha256.Sum256([]byte(raw))
-	return genericSource{
-		kind: genericSourceRelease, safe: "release://" + hex.EncodeToString(checksum[:]),
-		artifactURL: raw, revision: hex.EncodeToString(checksum[:]),
-	}, nil
-}
-
-func parseGenericHFSource(raw string) (genericSource, error) {
-	rest := strings.TrimSpace(raw[len("hf://"):])
-	if rest == "" || strings.ContainsAny(rest, "\x00?#\\") {
-		return genericSource{}, models.ErrModelReferenceInvalid
-	}
-	at := strings.LastIndex(rest, "@")
-	base, revision := rest, ""
-	if at >= 0 {
-		base, revision = rest[:at], rest[at+1:]
-	}
-	parts := strings.Split(base, "/")
-	if len(parts) < 2 || strings.TrimSpace(parts[0]) == "" || strings.TrimSpace(parts[1]) == "" {
-		return genericSource{}, models.ErrModelReferenceInvalid
-	}
-	for _, part := range parts {
-		if strings.TrimSpace(part) == "" || part == "." || part == ".." || strings.ContainsAny(part, " @\t\r\n") {
-			return genericSource{}, models.ErrModelReferenceInvalid
-		}
-	}
-	if strings.ContainsAny(revision, "\x00/@\\?# \t\r\n") {
-		return genericSource{}, models.ErrModelRevisionUnresolved
-	}
-	file := strings.Join(parts[2:], "/")
-	safe := "hf://" + parts[0] + "/" + parts[1]
-	if file != "" {
-		safe += "/" + file
-	}
-	if revision != "" {
-		safe += "@" + revision
-	}
-	return genericSource{
-		kind: genericSourceHF, safe: safe, owner: parts[0], repository: parts[1],
-		file: file, revision: revision,
-	}, nil
-}
-
-func genericHFSafeReference(source genericSource) string {
-	safe := "hf://" + source.owner + "/" + source.repository
-	if source.file != "" {
-		safe += "/" + source.file
-	}
-	return safe + "@" + source.revision
-}
-
-func genericRevisionFailure() error {
-	return &models.InvocationFailure{
-		Class:   models.InvocationFailureClassRevisionResolution,
-		Message: "model source revision could not be resolved to an immutable commit",
-		Cause:   models.ErrModelRevisionUnresolved,
-	}
-}
-
-func parseGenericFileSource(raw string) (genericSource, error) {
-	parsed, err := url.Parse(raw)
-	if err != nil || !validGenericFileURL(parsed) {
-		return genericSource{}, models.ErrModelReferenceInvalid
-	}
-	localPath, err := genericFileURLPath(parsed)
-	if err != nil {
-		return genericSource{}, models.ErrModelReferenceInvalid
-	}
-	if localPath == "" {
-		return genericSource{}, models.ErrModelReferenceInvalid
-	}
-	return genericSource{
-		kind: genericSourceFile, safe: "file://local", localPath: filepath.FromSlash(localPath),
-	}, nil
-}
-
-func validGenericFileURL(parsed *url.URL) bool {
-	return parsed.User == nil && parsed.RawQuery == "" && parsed.Fragment == ""
-}
-
-func genericFileURLPath(parsed *url.URL) (string, error) {
-	localPath, err := url.PathUnescape(parsed.Path)
-	if err != nil {
-		return "", err
-	}
-	if parsed.Host != "" && !strings.EqualFold(parsed.Host, "localhost") {
-		if len(parsed.Host) != 2 || parsed.Host[1] != ':' || !isASCIIAlphaByte(parsed.Host[0]) {
-			return "", models.ErrModelReferenceInvalid
-		}
-		localPath = parsed.Host + localPath
-	}
-	if len(localPath) >= 3 && localPath[0] == '/' && localPath[2] == ':' && isASCIIAlphaByte(localPath[1]) {
-		localPath = localPath[1:]
-	}
-	return localPath, nil
-}
-
-func isImmutableGenericRevision(value string) bool {
-	value = strings.TrimSpace(value)
-	if len(value) != 40 && len(value) != 64 {
-		return false
-	}
-	_, err := hex.DecodeString(value)
-	return err == nil
-}
-
-func (s *service) genericAssetURL(source genericSource, name string) string {
-	if source.kind == genericSourceRelease {
-		return source.artifactURL
-	}
-	return strings.TrimRight(s.endpoints.BaseURL, "/") + "/" + source.owner + "/" +
-		source.repository + "/resolve/" + source.revision + "/" + url.PathEscape(name) + "?download=true"
-}
-
-func (s *service) addGenericURLs(source genericSource, artifacts []genericArtifact) {
-	if source.kind != genericSourceHF {
-		return
-	}
-	for index := range artifacts {
-		if artifacts[index].url == "" {
-			artifacts[index].url = s.genericAssetURL(source, artifacts[index].requirement.Name)
-		}
-	}
-}
-
-func sourceDisplayName(source genericSource) string {
-	if source.kind == genericSourceHF {
-		return source.owner + "/" + source.repository
-	}
-	return "local-model"
-}
-
-func sourceMetadata(source genericSource) models.SourceMetadata {
-	if source.kind == genericSourceHF {
-		return models.SourceMetadata{Provider: "HUGGINGFACE", Reference: source.owner + "/" + source.repository, Revision: source.revision}
-	}
-	if source.kind == genericSourceRelease {
-		return models.SourceMetadata{Provider: "PINNED_BACKEND", Reference: "pinned-backend", Revision: source.revision}
-	}
-	return models.SourceMetadata{Provider: "LOCAL", Reference: source.safe}
-}
-
-func genericCacheKey(kind string, source genericSource, artifacts []genericArtifact) string {
-	names := genericArtifactIdentityNames(artifacts)
-	sort.Strings(names)
-	return kind + "|" + genericSourceIdentity(source) + "|" + strings.Join(names, ",")
-}
-
-func genericSourceIdentity(source genericSource) string {
-	if source.kind != genericSourceLocal && source.kind != genericSourceFile {
-		return source.safe
-	}
-	checksum := sha256.Sum256([]byte(source.localPath))
-	return source.safe + "|" + hex.EncodeToString(checksum[:])
-}
-
-func genericArtifactIdentityNames(artifacts []genericArtifact) []string {
-	names := make([]string, 0, len(artifacts))
-	for _, artifact := range artifacts {
-		requirement := artifact.requirement
-		names = append(names, fmt.Sprintf(
-			"%s:%d:%s",
-			requirement.Name,
-			requirement.Bytes,
-			strings.ToLower(strings.TrimSpace(requirement.SHA256)),
-		))
-	}
-	return names
-}
-
-func genericArtifactIdentityHash(kind string, source genericSource, artifacts []genericArtifact) string {
-	return genericIdentityHash(kind, source, genericArtifactIdentityNames(artifacts))
-}
-
-func genericIdentityHash(kind string, source genericSource, names []string) string {
-	identity := kind + "|" + genericSourceIdentity(source)
-	if len(names) > 0 {
-		cloned := append([]string(nil), names...)
-		sort.Strings(cloned)
-		identity += "|" + strings.Join(cloned, ",")
-	}
-	hash := sha256.Sum256([]byte(identity))
-	return hex.EncodeToString(hash[:])
-}
-
-func missingArtifactNames(artifacts []genericArtifact) []string {
-	names := make([]string, 0, len(artifacts))
-	for _, artifact := range artifacts {
-		names = append(names, artifact.requirement.Name)
-	}
-	sort.Strings(names)
-	return uniqueStrings(names)
-}
-
-func uniqueStrings(values []string) []string {
-	if len(values) == 0 {
-		return nil
-	}
-	result := make([]string, 0, len(values))
-	for _, value := range values {
-		if len(result) == 0 || result[len(result)-1] != value {
-			result = append(result, value)
-		}
-	}
-	return result
-}
-
-func isASCIIAlphaByte(value byte) bool {
-	return value >= 'a' && value <= 'z' || value >= 'A' && value <= 'Z'
-}
-
-func looksLikeLocalPath(value string) bool {
-	return filepath.IsAbs(value) || strings.HasPrefix(value, "./") || strings.HasPrefix(value, "../") ||
-		strings.HasPrefix(value, ".\\") || strings.HasPrefix(value, "..\\") ||
-		strings.ContainsAny(value, `/\\`) || filepath.Ext(value) != ""
 }
