+++ pkg/services/models/transports/cli/models_test.go
@@ -987,38 +987,3 @@ func TestManagedRuntimePullResponseErrorPreservesOutcomeDetails(t *testing.T) {
 		t.Fatalf("diagnostics = %q, want repository, operation, and status", diagnostics)
 	}
 }
-
-func TestModelsVerboseLogsFailureStatus(t *testing.T) {
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusNotFound)
-		_, _ = io.WriteString(w, `{"message":"model not found","family":"NOT_FOUND","code":"NOT_FOUND"}`)
-	}))
-	defer server.Close()
-
-	var diagnostics bytes.Buffer
-	_, err := queryModel(queryOptions{
-		Context:     context.Background(),
-		HTTP:        testHTTPProtocol(t),
-		Server:      strings.TrimSuffix(server.URL, "/"),
-		ModelName:   "missing",
-		Verbose:     true,
-		Diagnostics: &diagnostics,
-	})
-	if !errors.Is(err, ErrModelNotFound) {
-		t.Fatalf("queryModel error = %v, want ErrModelNotFound", err)
-	}
-	assertDiagnosticsContains(t, diagnostics.String(), []string{
-		"models inspect response",
-		"endpointPath=/models/missing",
-		"status=404",
-	})
-}
-
-func assertDiagnosticsContains(t *testing.T, got string, wants []string) {
-	t.Helper()
-	for _, want := range wants {
-		if !strings.Contains(got, want) {
-			t.Fatalf("diagnostics missing %q:\n%s", want, got)
-		}
-	}
-}
